requirements.txt
pandas==2.1.1
mysql-connector-python==8.1.0
xmltodict==0.13.0
python-dotenv==1.0.0
pytz==2023.3.post1
